+++ CHANGES.md
@@ -12,6 +12,7 @@ Version 1.1.0 (pending)
 - Added payment information to order detail screen
 - Add configuration option to require staff photo and signature
 - Add checks to ensure authorization number is 10 characters or less.
+- Added Staff Checked In by Department report with chart
 
 Version 1.0.1 (11/14/2019)
 - Staff check in: don't search staff list until 2 characters have been entered in search (for performance)